+++ resolver/similarity/__init__.py
@@ -1,3 +1,4 @@
 
 from .generic import ExactMatch
-from .vectors import CosineSimilarity, EuclideanDistance, JaccardSimilarity
+from .vectors import CosineSimilarity, EuclideanDistance
+from .sets import JaccardSimilarity
